requirements.txt
requests
python-dotenv
pandas
finta
pytest
flask
